Time: 2 ms (41.56%), Space: 113.5 MB (96.97%) - LeetHub

+++ 0643-maximum-average-subarray-i/0643-maximum-average-subarray-i.cpp
@@ -10,6 +10,6 @@ public:
             int sum = nums[i]-nums[i-k];
             maxEl = std::max(maxEl,sum);
         }
-        return (double(maxEl)/double(k));
+        return (double(maxEl)/k);
     }
 };
